+++ drivers/scsi/qla4xxx/ql4_def.h
@@ -732,7 +732,7 @@ struct scsi_qla_host {
 
 	/* qla82xx specific fields */
 	struct device_reg_82xx  __iomem *qla4_82xx_reg; /* Base I/O address */
-	unsigned long nx_pcibase;	/* Base I/O address */
+	uintptr_t nx_pcibase;		/* Base I/O address */
 	uint8_t *nx_db_rd_ptr;		/* Doorbell read pointer */
 	unsigned long nx_db_wr_ptr;	/* Door bell write pointer */
 	unsigned long first_page_group_start;
